chore: dissolve `time`
