+++ packages/shared/lib/core/network/helpers/getNetworkStatusFromNodeInfo.ts
@@ -1,4 +1,4 @@
-import { MILLISECONDS_PER_SECOND, SECONDS_PER_MINUTE } from '@lib/time'
+import { MILLISECONDS_PER_SECOND, SECONDS_PER_MINUTE } from '@core/utils'
 import { NETWORK_STATUS_DESCRIPTION } from '../constants'
 import { NetworkHealth } from '../enums'
 import { INetworkStatus } from '../interfaces'
