+++ packages/shared/lib/core/profile/constants/stronghold-password-clear-interval.constant.ts
@@ -1,3 +1,3 @@
-import { SECONDS_PER_MINUTE } from '@lib/time'
+import { SECONDS_PER_MINUTE } from '@core/utils'
 
 export const STRONGHOLD_PASSWORD_CLEAR_INTERVAL = 2 * SECONDS_PER_MINUTE
