+++ packages/shared/lib/core/profile/stores/timeSinceStrongholdUnlocked.store.ts
@@ -1,5 +1,7 @@
-import { time } from '@core/app/stores'
 import { derived } from 'svelte/store'
+
+import { time } from '@core/app/stores/time.store'
+
 import { timeStrongholdLastUnlocked } from './timeStrongholdLastUnlocked.store'
 
 export const timeSinceStrongholdUnlocked = derived(
