+++ packages/shared/lib/core/utils/constants/index.ts
@@ -3,3 +3,4 @@ export * from './hex.constants'
 export * from './iota-unit-map.constant'
 export * from './max-number-of-iotas.constant'
 export * from './pin-length.constant'
+export * from './time.constants'
