+++ packages/shared/lib/core/utils/constants/time.constants.ts
@@ -1,0 +1,9 @@
+export const MILLISECONDS_PER_SECOND = 1000
+export const SECONDS_PER_MINUTE = 60
+export const MINUTES_PER_HOUR = 60
+export const HOURS_PER_DAY = 24
+export const DAYS_PER_WEEK = 7
+export const MONTHS_PER_YEAR = 12
+
+// DERIVED
+export const SECONDS_PER_DAY = HOURS_PER_DAY * MINUTES_PER_HOUR * SECONDS_PER_MINUTE
