+++ packages/shared/lib/core/utils/types/duration.type.ts
@@ -1,0 +1,1 @@
+export type Duration = 'day' | 'hour' | 'minute' | 'second'
