+++ packages/shared/lib/core/utils/types/index.ts
@@ -1,2 +1,3 @@
+export * from './duration.type'
 export * from './iota-unit-map.type'
 export * from './values-of.type'
