+++ packages/shared/lib/core/wallet/actions/setOutgoingAsyncActivitiesToClaimed.ts
@@ -1,6 +1,6 @@
 import { IAccountState } from '@core/account'
 import { OutputData } from '@iota/wallet'
-import { MILLISECONDS_PER_SECOND } from '@lib/time'
+import { MILLISECONDS_PER_SECOND } from '@core/utils'
 import { get } from 'svelte/store'
 import { ActivityDirection, ActivityType } from '../enums'
 import { allAccountActivities, updateActivityDataByActivityId } from '../stores'
