+++ packages/shared/lib/core/wallet/utils/outputs/getTimelockDateFromOutput.ts
@@ -1,5 +1,5 @@
 import { OutputTypes } from '@iota/types'
-import { MILLISECONDS_PER_SECOND } from '@lib/time'
+import { MILLISECONDS_PER_SECOND } from '@core/utils'
 import { OUTPUT_TYPE_TREASURY, UNLOCK_CONDITION_TIMELOCK } from '../../constants'
 
 export function getTimelockDateFromOutput(output: OutputTypes): Date {
